+++ src/test/scala/uk/co/gresearch/spark/dgraph/graphframes/TestGraphFrames.scala
@@ -54,64 +54,64 @@ class TestGraphFrames extends FunSpec
 
       val expected = Set(
         Row(
-          Row(sw1, null, null, "Film", "Star Wars: Episode IV - A New Hope", Timestamp.valueOf("1977-05-25 00:00:00.0"), 7.75E8, 121, 0.8118081180811808.toFloat),
-          Row(sw1, leia, "starring", 0.25),
-          Row(leia, null, null, "Person", "Princess Leia", null, null, null, 1.3293357933579335.toFloat),
+          Row(dgraph.sw1, null, null, "Film", "Star Wars: Episode IV - A New Hope", Timestamp.valueOf("1977-05-25 00:00:00.0"), 7.75E8, 121, 0.8118081180811808.toFloat),
+          Row(dgraph.sw1, dgraph.leia, "starring", 0.25),
+          Row(dgraph.leia, null, null, "Person", "Princess Leia", null, null, null, 1.3293357933579335.toFloat),
         ),
         Row(
-          Row(sw2, null, null, "Film", "Star Wars: Episode V - The Empire Strikes Back", Timestamp.valueOf("1980-05-21 00:00:00.0"), 5.34E8, 124, 0.8118081180811808.toFloat),
-          Row(sw2, leia, "starring", 0.25),
-          Row(leia, null, null, "Person", "Princess Leia", null, null, null, 1.3293357933579335.toFloat),
+          Row(dgraph.sw2, null, null, "Film", "Star Wars: Episode V - The Empire Strikes Back", Timestamp.valueOf("1980-05-21 00:00:00.0"), 5.34E8, 124, 0.8118081180811808.toFloat),
+          Row(dgraph.sw2, dgraph.leia, "starring", 0.25),
+          Row(dgraph.leia, null, null, "Person", "Princess Leia", null, null, null, 1.3293357933579335.toFloat),
         ),
         Row(
-          Row(sw3, null, null, "Film", "Star Wars: Episode VI - Return of the Jedi", Timestamp.valueOf("1983-05-25 00:00:00.0"), 5.72E8, 131, 0.8118081180811808.toFloat),
-          Row(sw3, leia, "starring", 0.25),
-          Row(leia, null, null, "Person", "Princess Leia", null, null, null, 1.3293357933579335.toFloat),
+          Row(dgraph.sw3, null, null, "Film", "Star Wars: Episode VI - Return of the Jedi", Timestamp.valueOf("1983-05-25 00:00:00.0"), 5.72E8, 131, 0.8118081180811808.toFloat),
+          Row(dgraph.sw3, dgraph.leia, "starring", 0.25),
+          Row(dgraph.leia, null, null, "Person", "Princess Leia", null, null, null, 1.3293357933579335.toFloat),
         ),
         Row(
-          Row(sw2, null, null, "Film", "Star Wars: Episode V - The Empire Strikes Back", Timestamp.valueOf("1980-05-21 00:00:00.0"), 5.34E8, 124, 0.8118081180811808.toFloat),
-          Row(sw2, irvin, "director", 0.25),
-          Row(irvin, null, null, "Person", "Irvin Kernshner", null, null, null, 0.9843173431734319.toFloat),
+          Row(dgraph.sw2, null, null, "Film", "Star Wars: Episode V - The Empire Strikes Back", Timestamp.valueOf("1980-05-21 00:00:00.0"), 5.34E8, 124, 0.8118081180811808.toFloat),
+          Row(dgraph.sw2, dgraph.irvin, "director", 0.25),
+          Row(dgraph.irvin, null, null, "Person", "Irvin Kernshner", null, null, null, 0.9843173431734319.toFloat),
         ),
         Row(
-          Row(sw1, null, null, "Film", "Star Wars: Episode IV - A New Hope", Timestamp.valueOf("1977-05-25 00:00:00.0"), 7.75E8, 121, 0.8118081180811808.toFloat),
-          Row(sw1, han, "starring", 0.25),
-          Row(han, null, null, "Person", "Han Solo", null, null, null, 1.3293357933579335.toFloat),
+          Row(dgraph.sw1, null, null, "Film", "Star Wars: Episode IV - A New Hope", Timestamp.valueOf("1977-05-25 00:00:00.0"), 7.75E8, 121, 0.8118081180811808.toFloat),
+          Row(dgraph.sw1, dgraph.han, "starring", 0.25),
+          Row(dgraph.han, null, null, "Person", "Han Solo", null, null, null, 1.3293357933579335.toFloat),
         ),
         Row(
-          Row(sw2, null, null, "Film", "Star Wars: Episode V - The Empire Strikes Back", Timestamp.valueOf("1980-05-21 00:00:00.0"), 5.34E8, 124, 0.8118081180811808.toFloat),
-          Row(sw2, han, "starring", 0.25),
-          Row(han, null, null, "Person", "Han Solo", null, null, null, 1.3293357933579335.toFloat),
+          Row(dgraph.sw2, null, null, "Film", "Star Wars: Episode V - The Empire Strikes Back", Timestamp.valueOf("1980-05-21 00:00:00.0"), 5.34E8, 124, 0.8118081180811808.toFloat),
+          Row(dgraph.sw2, dgraph.han, "starring", 0.25),
+          Row(dgraph.han, null, null, "Person", "Han Solo", null, null, null, 1.3293357933579335.toFloat),
         ),
         Row(
-          Row(sw3, null, null, "Film", "Star Wars: Episode VI - Return of the Jedi", Timestamp.valueOf("1983-05-25 00:00:00.0"), 5.72E8, 131, 0.8118081180811808.toFloat),
-          Row(sw3, han, "starring", 0.25),
-          Row(han, null, null, "Person", "Han Solo", null, null, null, 1.3293357933579335.toFloat),
+          Row(dgraph.sw3, null, null, "Film", "Star Wars: Episode VI - Return of the Jedi", Timestamp.valueOf("1983-05-25 00:00:00.0"), 5.72E8, 131, 0.8118081180811808.toFloat),
+          Row(dgraph.sw3, dgraph.han, "starring", 0.25),
+          Row(dgraph.han, null, null, "Person", "Han Solo", null, null, null, 1.3293357933579335.toFloat),
         ),
         Row(
-          Row(sw1, null, null, "Film", "Star Wars: Episode IV - A New Hope", Timestamp.valueOf("1977-05-25 00:00:00.0"), 7.75E8, 121, 0.8118081180811808.toFloat),
-          Row(sw1, lucas, "director", 0.25),
-          Row(lucas, null, null, "Person", "George Lucas", null, null, null, 0.9843173431734319.toFloat),
+          Row(dgraph.sw1, null, null, "Film", "Star Wars: Episode IV - A New Hope", Timestamp.valueOf("1977-05-25 00:00:00.0"), 7.75E8, 121, 0.8118081180811808.toFloat),
+          Row(dgraph.sw1, dgraph.lucas, "director", 0.25),
+          Row(dgraph.lucas, null, null, "Person", "George Lucas", null, null, null, 0.9843173431734319.toFloat),
         ),
         Row(
-          Row(sw1, null, null, "Film", "Star Wars: Episode IV - A New Hope", Timestamp.valueOf("1977-05-25 00:00:00.0"), 7.75E8, 121, 0.8118081180811808.toFloat),
-          Row(sw1, luke, "starring", 0.25),
-          Row(luke, null, null, "Person", "Luke Skywalker", null, null, null, 1.3293357933579335.toFloat),
+          Row(dgraph.sw1, null, null, "Film", "Star Wars: Episode IV - A New Hope", Timestamp.valueOf("1977-05-25 00:00:00.0"), 7.75E8, 121, 0.8118081180811808.toFloat),
+          Row(dgraph.sw1, dgraph.luke, "starring", 0.25),
+          Row(dgraph.luke, null, null, "Person", "Luke Skywalker", null, null, null, 1.3293357933579335.toFloat),
         ),
         Row(
-          Row(sw2, null, null, "Film", "Star Wars: Episode V - The Empire Strikes Back", Timestamp.valueOf("1980-05-21 00:00:00.0"), 5.34E8, 124, 0.8118081180811808.toFloat),
-          Row(sw2, luke, "starring", 0.25),
-          Row(luke, null, null, "Person", "Luke Skywalker", null, null, null, 1.3293357933579335.toFloat),
+          Row(dgraph.sw2, null, null, "Film", "Star Wars: Episode V - The Empire Strikes Back", Timestamp.valueOf("1980-05-21 00:00:00.0"), 5.34E8, 124, 0.8118081180811808.toFloat),
+          Row(dgraph.sw2, dgraph.luke, "starring", 0.25),
+          Row(dgraph.luke, null, null, "Person", "Luke Skywalker", null, null, null, 1.3293357933579335.toFloat),
         ),
         Row(
-          Row(sw3, null, null, "Film", "Star Wars: Episode VI - Return of the Jedi", Timestamp.valueOf("1983-05-25 00:00:00.0"), 5.72E8, 131, 0.8118081180811808.toFloat),
-          Row(sw3, luke, "starring", 0.25),
-          Row(luke, null, null, "Person", "Luke Skywalker", null, null, null, 1.3293357933579335.toFloat),
+          Row(dgraph.sw3, null, null, "Film", "Star Wars: Episode VI - Return of the Jedi", Timestamp.valueOf("1983-05-25 00:00:00.0"), 5.72E8, 131, 0.8118081180811808.toFloat),
+          Row(dgraph.sw3, dgraph.luke, "starring", 0.25),
+          Row(dgraph.luke, null, null, "Person", "Luke Skywalker", null, null, null, 1.3293357933579335.toFloat),
         ),
         Row(
-          Row(sw3, null, null, "Film", "Star Wars: Episode VI - Return of the Jedi", Timestamp.valueOf("1983-05-25 00:00:00.0"), 5.72E8, 131, 0.8118081180811808.toFloat),
-          Row(sw3, richard, "director", 0.25),
-          Row(richard, null, null, "Person", "Richard Marquand", null, null, null, 0.9843173431734319.toFloat),
+          Row(dgraph.sw3, null, null, "Film", "Star Wars: Episode VI - Return of the Jedi", Timestamp.valueOf("1983-05-25 00:00:00.0"), 5.72E8, 131, 0.8118081180811808.toFloat),
+          Row(dgraph.sw3, dgraph.richard, "director", 0.25),
+          Row(dgraph.richard, null, null, "Person", "Richard Marquand", null, null, null, 0.9843173431734319.toFloat),
         )
       )
       assert(triplets === expected)
@@ -120,17 +120,17 @@ class TestGraphFrames extends FunSpec
     def doVerticesTest(load: () => DataFrame): Unit = {
       val vertices = load().collect().toSet
       val expected = Set(
-        Row(graphQlSchema, "", "dgraph.graphql.schema", "dgraph.graphql", null, null, null, null),
-        Row(st1, null, null, "Film", "Star Trek: The Motion Picture", Timestamp.valueOf("1979-12-07 00:00:00.0"), 1.39E8, 132),
-        Row(leia, null, null, "Person", "Princess Leia", null, null, null),
-        Row(lucas, null, null, "Person", "George Lucas", null, null, null),
-        Row(irvin, null, null, "Person", "Irvin Kernshner", null, null, null),
-        Row(sw1, null, null, "Film", "Star Wars: Episode IV - A New Hope", Timestamp.valueOf("1977-05-25 00:00:00.0"), 7.75E8, 121),
-        Row(sw2, null, null, "Film", "Star Wars: Episode V - The Empire Strikes Back", Timestamp.valueOf("1980-05-21 00:00:00.0"), 5.34E8, 124),
-        Row(luke, null, null, "Person", "Luke Skywalker", null, null, null),
-        Row(han, null, null, "Person", "Han Solo", null, null, null),
-        Row(richard, null, null, "Person", "Richard Marquand", null, null, null),
-        Row(sw3, null, null, "Film", "Star Wars: Episode VI - Return of the Jedi", Timestamp.valueOf("1983-05-25 00:00:00.0"), 5.72E8, 131)
+        Row(dgraph.graphQlSchema, "", "dgraph.graphql.schema", "dgraph.graphql", null, null, null, null),
+        Row(dgraph.st1, null, null, "Film", "Star Trek: The Motion Picture", Timestamp.valueOf("1979-12-07 00:00:00.0"), 1.39E8, 132),
+        Row(dgraph.leia, null, null, "Person", "Princess Leia", null, null, null),
+        Row(dgraph.lucas, null, null, "Person", "George Lucas", null, null, null),
+        Row(dgraph.irvin, null, null, "Person", "Irvin Kernshner", null, null, null),
+        Row(dgraph.sw1, null, null, "Film", "Star Wars: Episode IV - A New Hope", Timestamp.valueOf("1977-05-25 00:00:00.0"), 7.75E8, 121),
+        Row(dgraph.sw2, null, null, "Film", "Star Wars: Episode V - The Empire Strikes Back", Timestamp.valueOf("1980-05-21 00:00:00.0"), 5.34E8, 124),
+        Row(dgraph.luke, null, null, "Person", "Luke Skywalker", null, null, null),
+        Row(dgraph.han, null, null, "Person", "Han Solo", null, null, null),
+        Row(dgraph.richard, null, null, "Person", "Richard Marquand", null, null, null),
+        Row(dgraph.sw3, null, null, "Film", "Star Wars: Episode VI - Return of the Jedi", Timestamp.valueOf("1983-05-25 00:00:00.0"), 5.72E8, 131)
       )
       assert(vertices === expected)
     }
@@ -138,25 +138,25 @@ class TestGraphFrames extends FunSpec
     def doEdgesTest(load: () => DataFrame): Unit = {
       val edges = load().collect().toSet
       val expected = Set(
-        Row(sw1, leia, "starring"),
-        Row(sw1, lucas, "director"),
-        Row(sw1, luke, "starring"),
-        Row(sw1, han, "starring"),
-        Row(sw2, leia, "starring"),
-        Row(sw2, irvin, "director"),
-        Row(sw2, luke, "starring"),
-        Row(sw2, han, "starring"),
-        Row(sw3, leia, "starring"),
-        Row(sw3, luke, "starring"),
-        Row(sw3, han, "starring"),
-        Row(sw3, richard, "director")
+        Row(dgraph.sw1, dgraph.leia, "starring"),
+        Row(dgraph.sw1, dgraph.lucas, "director"),
+        Row(dgraph.sw1, dgraph.luke, "starring"),
+        Row(dgraph.sw1, dgraph.han, "starring"),
+        Row(dgraph.sw2, dgraph.leia, "starring"),
+        Row(dgraph.sw2, dgraph.irvin, "director"),
+        Row(dgraph.sw2, dgraph.luke, "starring"),
+        Row(dgraph.sw2, dgraph.han, "starring"),
+        Row(dgraph.sw3, dgraph.leia, "starring"),
+        Row(dgraph.sw3, dgraph.luke, "starring"),
+        Row(dgraph.sw3, dgraph.han, "starring"),
+        Row(dgraph.sw3, dgraph.richard, "director")
       )
       assert(edges === expected)
     }
 
     Seq(
-      ("target", () => Seq(cluster.grpc)),
-      ("targets", () => Seq(cluster.grpc, cluster.grpcLocalIp))
+      ("target", () => Seq(dgraph.target)),
+      ("targets", () => Seq(dgraph.target, dgraph.targetLocalIp))
     ).foreach{case (test, targets) =>
 
       it(s"should load dgraph from $test via implicit session") {
@@ -164,7 +164,7 @@ class TestGraphFrames extends FunSpec
       }
 
       it(s"should load dgraph from $test via reader") {
-        doGraphTest(() => spark.read.dgraph(targets(): _*))
+        doGraphTest(() => spark.read.dgraph.graphframes(targets(): _*))
       }
 
       it(s"should load vertices from $test via implicit session") {
@@ -172,7 +172,7 @@ class TestGraphFrames extends FunSpec
       }
 
       it(s"should load vertices from $test via reader") {
-        doVerticesTest(() => spark.read.dgraphVertices(targets(): _*))
+        doVerticesTest(() => spark.read.dgraph.vertices(targets(): _*))
       }
 
       it(s"should load edges from $test via implicit session") {
@@ -180,7 +180,7 @@ class TestGraphFrames extends FunSpec
       }
 
       it(s"should load edges from $test via reader") {
-        doEdgesTest(() => spark.read.dgraphEdges(targets(): _*))
+        doEdgesTest(() => spark.read.dgraph.edges(targets(): _*))
       }
 
     }